spike: namespace expert codebase
By namespacing the Expert applicaiton as well, we can seamlessly
communcate with the Engine application without having have knowledge of
the namespaced module names.

Doing this naively proved to come with some tradeoffs. The namespacing
as it was before would also namespace the application files. This causes
the application config to get a little hairy. Lexical will also
namespace the config files, but that works because they seem to do this
when "packaging", which copies the files into a release like structure.

I wanted to be able to run the project as "vanilla" as possible, so to
do so, a modification to the namespace mix task allows us to namespace
with or without changing the applications. Namespacing the applications
is not necessary for the Expert codebase, as the namespacing is done to
allow seamless communciation with the engine, and not to avoid
collisions with user code.

This works nicely, but since we are namespacing the beam files in place,
when we make a change and recompile, mix recompiles everything from
scratch.

To avoid doing this, tooling in the justfile (just is analagous to make)
will swap the original compilation artifiacts in and out with the
namespaced ones so mix will continue to incrementaly compile.

It will namespace everything rather than incrementally namespace, but
this happens fast enought that it doesn't seem to be a problem as of
now.