spike(namespace): parameterize transformers
This patch primarily parameterizes the transformers to accept the apps
and the root modules they need to work on as parameters instead as
global state.

This increases the flexibility of the project by hoisting this
configurability to the boundary, allowing it to be overridden/customized
by CLI options.

This was accomplished by passing an "opts" parameter to each transformer
function (which were consequently renamed from "apply" to "run" to avoid
a naming collision with Kernel.apply/2).

Incidental changes are moving the transformers from the mix task
directory to the regular lib directory and to add unit tests to ensure
that there weren't any regressions.

Something to note, I don't think that the `configs`, `scripts`, or
`boots` transformers are needed any longer, but that conclusion should
be double checked.