spike(tests): add integration tests for expert
This patch introduces integration tests using GenLSPs builtin testing
sdk.

The test sdk starts an in process instance of the server and
communicates with it via the TCP adapter. In tests, the test is the
"client" instead of a text editor.

Tests however are not namespaced. From what I could see, tests in
Lexical don't seem to be namespaced. Attempting to do so I believe would
be a rabbit hole, as you'd have to namespace the in memory compiled
tests before they run.

---

This also adjusts how the project is namespaced in development.
Previously, we would copy the compiled beam files to another directory
and namespace the ones in the standard one, then copy the normal files
back before compiling again to achieve normal incremental compiles.

Now, it leaves the normal files in place, but copies them to another
directory and namespaces them there. Then when we start the server, we
change the `MIX_BUILD_PATH` to the new directory.

---

One last thing to note is that the `:engine` atom is not properly
namespaced in the Expert project, as we aren't namespacing applications.
But, it was attempting to start the Engine application via
`Application.ensure_all_started(:engine)`. To get around this, I put a
function in the Engine project that calls that instead, and the Expert
projects RPC call is `Engine.ensure_all_started()`.