Add .tool-versions
The README.md suggests to use those versions, and the pre-commit hooks
depends on the formatting done by that elixir version. Having a
`.tool-versions` file helps making sure the right versions are being
used if the developer is using a version manager like asdf. Otherwise
the global elixir and otp versions will be used, which changed formatting
rules after 1.13.4.