chore: silence spammy log messages (#40)
* chore: Silence spammy log messages

For a while now, elixir_sense has been emitting lots of log messages
that spammed our logs and made it harder to see what was going
on. This PR removes the spam.

Now, we remove any log message that has "is already compiled." in it,
which handles all of the messages caused by elixir_sense.

* fix: remove unused argument

---------

Co-authored-by: Dorgan <dorgandash@gmail.com>